pyproject.toml
[tool.poetry]
name = "event-ticket-service"
version = "0.1.0"
description = ""
authors = ["Jinay Shah <jinayshah86@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.3"
flask-sqlalchemy = "^3.0.3"
flask-migrate = "^4.0.4"
flask-cors = "^3.0.10"
flask-marshmallow = "^0.15.0"
marshmallow-sqlalchemy = "^0.29.0"
python-dotenv = "^1.0.0"
flask-restful = "^0.3.9"
psycopg2-binary = "^2.9.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
